+++ client/src/app/api/university/[university]/department/[department]/route.ts
@@ -1,5 +1,6 @@
 import { supabase } from '@/lib/dbConnect';
 import { NextResponse } from 'next/server';
+import { DELETE as del } from './branch/[branch]/route';
 
 export const PUT = async(req:any, res:any)=>{
     const department = await req.json();
@@ -25,6 +26,28 @@ export const DELETE = async(req:any, res:any)=>{
     const id = req.url!.split("department/")[1]
 
     try {
+        const { data: branchData, error: branchError } = await supabase
+            .from('branch')
+            .select('id')
+            .eq('dept_id', id);
+
+        if (branchError) {
+            throw branchError;
+        }
+        
+        for (const branchItem of branchData) {
+            await del({ url: `/branch/${branchItem.id}` },0);
+        }
+
+        const { data:session_data, error:session_error } = await supabase
+            .from('session')
+            .delete()
+            .eq('dept_id', id);
+            
+        if (session_error) {
+            throw session_error;
+        }
+
         const { data, error } = await supabase
             .from('department')
             .delete()
@@ -36,5 +59,24 @@ export const DELETE = async(req:any, res:any)=>{
     } catch (error:any) {
         console.error(error);
         return NextResponse.json({status: 500, error_message: error.message, function_name: 'delete_department' });
+    }
+}
+
+
+export const GET = async(req:any, res:any)=>{
+    const id = req.url!.split("department/")[1]
+
+    try {
+        const { data, error } = await supabase
+            .from('department')
+            .select()
+            .eq('id', id)
+        if (error) {
+            throw error
+        }
+        return NextResponse.json({status: 201, data: data, function_name: 'get_department'})
+    } catch (error:any) {
+        console.error(error)
+        return NextResponse.json({status: 500, error_message: error.message, function_name: 'get_department'})
     }
 }
